fix data layer to match dataloader requirements
Batch functions output should match input length, and the output
order should match the input order.

For example, for the input `fn(1, 2)`, the output must be be
`[{id: 1, ...}, {id: 2, ...}]`.
To solve this problem I created `sortByIdsOrder(ids)(rows)`.

The other problem was the inner join. It must be a left join,
because products not related to any store still need to be
returned by the query, with an empty array of storeIds.

Ps.: ARRAY_AGG returns [null] instead of [] when no stores
relates to the product, so I'm using ARRAY_REMOVE(arr, NULL)
to fix that.