change relation strategy focusing on pagination/filters/etc
The old strategy was to just simply eager load relation ids on
parent resolver. That doesn't work when you want to paginate
relations.

The new strategy doesnt eager load those relations, instead it
relies on loader composition: a relation loader that uses the
related model loader. (For example, loaders.storesByProductId
that uses loaders.storeById).

So, first there is the relation loader. At first it didn't thought
that we need loaders for relations, but then I realized that this:
query {
  stores {
    products {
      stores {
        products {
          name
        }
      }
    }
  }
}
That's not a realistic query, but it's something that can happen.
If we don't cache the relation itself, we are going to let a DoS
door open: a very deep query easy to craft is going to take tons
of the server to resolve.

So, as soon as we have at least this situation, it's possible that
other complex queries can make our resolvers to hit the data layer
for the same relation between two same objects.

Second, there is the specific model loader that's used by the
relation loader. The relation loader only fetchs the data layer
for relation ids, and calls the specific model loader to load the
models contents. The idea here is that sometimes the model was
already loaded, so we can fetch the data layer only for the ids.

For example:
1) We call storesByProductId.load(1001) on Product.stores resolver
2) storesByProductId loader fetchs only ids on data layer calling
getStoresByProductsIds. This is the only data layer call that will
always happen on this loading.
3) Now we can load store contents using storeById loader. A store
store that was already loaded has its contents on the cache, no
need to fetch it from database again.